add support for redlock

+++ tests/common/test_distributed.py
@@ -11,18 +11,21 @@ from multiprocessing import Event, Process, Queue
 from pathlib import Path
 from typing import (
     Any,
+    Callable,
     Iterable,
     List,
+    Tuple,
 )
 
 import attr
 from etcetra.types import HostPortPair as EtcdHostPortPair
 import pytest
+from redis.asyncio import Redis
 
 from ai.backend.common.distributed import GlobalTimer
 from ai.backend.common.events import AbstractEvent, EventDispatcher, EventProducer
-from ai.backend.common.lock import EtcdLock, FileLock
-from ai.backend.common.types import AgentId, EtcdRedisConfig, HostPortPair
+from ai.backend.common.lock import AbstractDistributedLock, EtcdLock, FileLock, RedisLock
+from ai.backend.common.types import AgentId, EtcdRedisConfig, HostPortPair, RedisConnectionInfo
 
 from ai.backend.common.etcd import AsyncEtcd, ConfigScopes
 
@@ -70,18 +73,55 @@ class NoopEvent(AbstractEvent):
         return cls(value[0])
 
 
+async def run_timer(
+    lock_factory: Callable[[None], AbstractDistributedLock],
+    stop_event: asyncio.Event,
+    event_records: List[float],
+    redis_addr: RedisConnectionInfo,
+    test_ns: str,
+    interval: int,
+) -> None:
+    async def _tick(context: Any, source: AgentId, event: NoopEvent) -> None:
+        print("_tick")
+        event_records.append(time.monotonic())
+
+    redis_config = EtcdRedisConfig(addr=redis_addr)
+    event_dispatcher = await EventDispatcher.new(
+        redis_config,
+        node_id=test_ns,
+    )
+    event_producer = await EventProducer.new(
+        redis_config,
+    )
+    event_dispatcher.consume(NoopEvent, None, _tick)
+
+    timer = GlobalTimer(
+        lock_factory(),
+        event_producer,
+        lambda: NoopEvent(test_ns),
+        interval,
+    )
+    try:
+        await timer.join()
+        await stop_event.wait()
+    finally:
+        await timer.leave()
+        await event_producer.close()
+        await event_dispatcher.close()
+
+
 class TimerNode(threading.Thread):
 
     def __init__(
         self,
         event_records: list[float],
-        lock_path: Path,
+        lock_factory: Callable[[None], AbstractDistributedLock],
         thread_idx: int,
         timer_ctx: TimerNodeContext,
     ) -> None:
         super().__init__()
         self.event_records = event_records
-        self.lock_path = lock_path
+        self.lock_factory = lock_factory
         self.thread_idx = thread_idx
         self.interval = timer_ctx.interval
         self.test_ns = timer_ctx.test_ns
@@ -106,7 +146,7 @@ class TimerNode(threading.Thread):
         event_dispatcher.consume(NoopEvent, None, _tick)
 
         timer = GlobalTimer(
-            FileLock(self.lock_path, timeout=0, debug=True),
+            self.lock_factory(),
             event_producer,
             lambda: NoopEvent(self.test_ns),
             self.interval,
@@ -124,9 +164,24 @@ class TimerNode(threading.Thread):
 
 
 @pytest.mark.asyncio
-async def test_global_timer_filelock(request, test_ns, redis_container) -> None:
-    lock_path = Path(tempfile.gettempdir()) / f'{test_ns}.lock'
-    request.addfinalizer(partial(lock_path.unlink, missing_ok=True))
+@pytest.mark.parametrize("lock_type", ['file', 'redis'])
+async def test_global_timer(request, test_ns, redis_container, lock_type) -> None:
+    if lock_type == 'file':
+        lock_path = Path(tempfile.gettempdir()) / f'{test_ns}.lock'
+        print('lock path', lock_path)
+        request.addfinalizer(partial(lock_path.unlink, missing_ok=True))
+        lock_factory = lambda: FileLock(lock_path, timeout=0, debug=True)
+    elif lock_type == 'redis':
+        addr = redis_container[1]
+        print('redis addr', addr)
+        r = RedisConnectionInfo(
+            Redis.from_url(url=f'redis://{addr.host}:{addr.port}', socket_timeout=0.5),
+            service_name=None,
+        )
+        assert isinstance(r.client, Redis)
+        lock_factory = lambda: RedisLock(f'{test_ns}lock', r.client, debug=True)
+    else:
+        pytest.fail('should not reach here')
     event_records: List[float] = []
     num_threads = 7
     num_records = 0
@@ -137,7 +192,7 @@ async def test_global_timer_filelock(request, test_ns, redis_container) -> None:
     for thread_idx in range(num_threads):
         timer_node = TimerNode(
             event_records,
-            lock_path,
+            lock_factory,
             thread_idx,
             TimerNodeContext(
                 test_ns=test_ns,
@@ -157,6 +212,46 @@ async def test_global_timer_filelock(request, test_ns, redis_container) -> None:
     print("joining timer threads")
     for timer_node in threads:
         timer_node.join()
+    print("checking records")
+    print(event_records)
+    num_records = len(event_records)
+    print(f"{num_records=}")
+    assert target_count - 2 <= num_records <= target_count + 2
+
+
+@pytest.mark.asyncio
+async def test_gloal_timer_task(test_ns, redis_container) -> None:
+    redis_addr = redis_container[1]
+    r = RedisConnectionInfo(
+        Redis.from_url(f'redis://{redis_addr.host}:{redis_addr.port}'),
+        service_name=None,
+    )
+    assert isinstance(r.client, Redis)
+    lock_factory = lambda: RedisLock(f'{test_ns}lock', r.client, debug=True)
+
+    event_records: List[float] = []
+    num_threads = 7
+    num_records = 0
+    delay = 3.0
+    interval = 0.5
+    target_count = (delay / interval)
+    tasks: List[Tuple[asyncio.Task, asyncio.Event]] = []
+    for thread_idx in range(num_threads):
+        stop_event = asyncio.Event()
+        task = asyncio.create_task(
+            run_timer(lock_factory, stop_event, event_records, redis_addr, test_ns, interval)
+        )
+        tasks.append((task, stop_event))
+    print(f"spawned {num_threads} timers")
+    print(tasks)
+    print("waiting")
+    await asyncio.sleep(delay)
+    print("stopping timers")
+    for _, stop_event in tasks:
+        stop_event.set()
+    print("joining timer tasks")
+    for timer_task, _ in tasks:
+        await timer_task
     print("checking records")
     print(event_records)
     num_records = len(event_records)
